Refactoring: isolate listener bookkepping in own class
